test: run OAuth Slack resume through real agent
