+++ scripts/test-architecture-debt.json
@@ -3,7 +3,6 @@
     "packages/junior/tests/integration/local-agent-runner.test.ts": 18,
     "packages/junior/tests/integration/mcp-oauth-callback.test.ts": 2,
     "packages/junior/tests/integration/oauth-callback.test.ts": 3,
-    "packages/junior/tests/integration/oauth-resume-slack.test.ts": 4,
     "packages/junior/tests/integration/slack/assistant-thread-contract.test.ts": 2,
     "packages/junior/tests/integration/slack/attachment-behavior.test.ts": 3,
     "packages/junior/tests/integration/slack/attachment-media-behavior.test.ts": 4,
@@ -23,7 +22,6 @@
   "scripted-agent-runner": {
     "packages/junior/tests/integration/api-turn-work.test.ts": 4,
     "packages/junior/tests/integration/local-agent-runner.test.ts": 3,
-    "packages/junior/tests/integration/oauth-resume-slack.test.ts": 6,
     "packages/junior/tests/integration/slack/finalized-reply-behavior.test.ts": 7
   }
 }
